Restructure component folders
Nest components in their own folders, add index.js per component.
