+++ components/molecules/Alert/index.js
@@ -1,0 +1,1 @@
+export {Info, Warning} from './Alert.js'
