+++ components/molecules/Hero/index.js
@@ -1,0 +1,1 @@
+export {Hero} from './Hero'
