+++ pages/isr.js
@@ -1,8 +1,8 @@
-import PropTypes from 'prop-types'
 import Layout from '@/components/common/Layout'
+import {Info} from '@/components/molecules/Alert'
 import {fetcher} from '@/lib/functions'
 import Link from 'next/link'
-import {Info} from '@/components/molecules/Alerts'
+import PropTypes from 'prop-types'
 
 export default function IncrementalStaticRegeneration(props) {
   return (
