+++ pages/posts/[id].js
@@ -1,7 +1,7 @@
-import PropTypes from 'prop-types'
 import Layout from '@/components/common/Layout'
+import {Info} from '@/components/molecules/Alert'
 import {fetcher} from '@/lib/functions'
-import {Info} from '@/components/molecules/Alerts'
+import PropTypes from 'prop-types'
 
 export default function BlogPost({post}) {
   return (
